moves cli specs in separate files
